# COLLEGE/iiitv/SEM_3/Web_Services/social_media/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NamoGram RESTful Web Services API
  version: 1.0.0
  description: |
    Comprehensive RESTful API for NamoGram (Instagram-like social media platform with modern blue aesthetic).
    Fulfills all Module 2 REST requirements:
    - Resource-shaped URLs under /api/v1
    - Correct HTTP methods (GET, POST, PUT, PATCH, DELETE)
    - 8+ distinct status codes (200, 201, 204, 304, 400, 401, 403, 404, 409, 422, 429, 503)
    - One consistent RFC error body
    - Idempotency-Key support on POST
    - ETag and 304 Not Modified
    - Deliberate Cache-Control headers
    - Ajv JSON Schema validation
    - Content negotiation (application/json and application/xml)
    - Cursor pagination with next links
    - Sliding window rate limiting with headers
servers:
  - url: http://localhost:5000
    description: Local Development Server
paths:
  /api/v1/auth/login:
    post:
      summary: Authenticate user and receive JWT token
      operationId: loginUser
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/auth/me:
    get:
      summary: Retrieve current authenticated user profile
      operationId: getCurrentUser
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/posts:
    get:
      summary: Retrieve posts with cursor pagination and cache control
      operationId: getPosts
      tags: [Posts]
      parameters:
        - name: limit
          in: query
          description: Max number of items to return (1-50)
          schema:
            type: integer
            default: 5
        - name: cursor
          in: query
          description: Post ID cursor for keyset pagination
          schema:
            type: string
        - name: tag
          in: query
          description: Filter posts by tag name
          schema:
            type: string
        - name: Accept
          in: header
          description: Content negotiation header (application/json or application/xml)
          schema:
            type: string
            default: application/json
      responses:
        '200':
          description: Paginated posts collection
          headers:
            ETag:
              schema: { type: string }
              description: Entity tag for conditional requests
            Cache-Control:
              schema: { type: string }
              description: Deliberate cache header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsFeedResponse'
            application/xml:
              schema:
                type: string
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      summary: Create a new post with schema validation and idempotency
      operationId: createPost
      tags: [Posts]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Unique client key to guarantee idempotent creation
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateInput'
      responses:
        '201':
          description: Post created successfully
          headers:
            Location:
              schema: { type: string }
              description: URI of newly created resource
            Idempotent-Replay:
              schema: { type: string }
              description: Returned if response is replayed from cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: melanie_v }
        password: { type: string, example: password123 }
    AuthResponse:
      type: object
      properties:
        message: { type: string }
        token: { type: string }
        user:
          $ref: '#/components/schemas/User'
    UserProfileResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string }
        avatar: { type: string }
        bio: { type: string }
        location: { type: string }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        verified: { type: boolean }
    PostCreateInput:
      type: object
      required: [caption, imageUrl]
      properties:
        caption: { type: string, minLength: 1 }
        imageUrl: { type: string }
        location: { type: string }
        eventTime: { type: string }
        capacity: { type: string }
        timeRemaining: { type: string }
        tags: { type: array, items: { type: string } }
...
    Post:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        author:
          $ref: '#/components/schemas/User'
        caption: { type: string }
        imageUrl: { type: string }
        location: { type: string }
        eventTime: { type: string }
        capacity: { type: string }
        timeRemaining: { type: string }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        isLiked: { type: boolean }
        isSaved: { type: boolean }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    PostsFeedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        pagination:
          type: object
          properties:
            limit: { type: integer }
            cursor: { type: string, nullable: true }
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
            count: { type: integer }
            total: { type: integer }
        links:
          type: object
          properties:
            self: { type: string }
            next: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp, correlationId]
          properties:
            code: { type: string, example: NOT_FOUND }
            message: { type: string, example: Resource was not found. }
            details: { nullable: true }
            timestamp: { type: string, format: date-time }
            correlationId: { type: string, example: 550e8400-e29b-41d4-a716-446655440000 }
  responses:
    NotModified:
      description: 304 Not Modified - Resource has not changed based on If-None-Match ETag
    BadRequestError:
      description: 400 Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: 401 Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: 422 Unprocessable Entity - JSON Schema validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: 429 Too Many Requests
      headers:
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer } }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'